native vs jgit
