+++ src/main/groovy/com/palantir/gradle/gitversion/GitDescribe.groovy
@@ -1,0 +1,6 @@
+package com.palantir.gradle.gitversion
+
+interface GitDescribe {
+
+    String describe(String prefix)
+}
