+++ src/main/groovy/com/palantir/gradle/gitversion/JGitDescribe.groovy
@@ -1,0 +1,15 @@
+package com.palantir.gradle.gitversion
+
+class JGitDescribe implements GitDescribe {
+
+    private File directory
+
+    JGitDescribe(File directory) {
+        this.directory = directory
+    }
+
+    @Override
+    String describe(String prefix) {
+        return null
+    }
+}
